Add fn TileExt::find_widget_rect
